+++ packages/agreement-process/test/createAgreement.test.ts
@@ -1,0 +1,645 @@
+import { fail } from "assert";
+import {
+  decodeProtobufPayload,
+  expectPastTimestamp,
+  getRandomAuthData,
+  getMockDescriptorPublished,
+  getMockEServiceAttribute,
+  getMockEService,
+  getMockTenant,
+  getMockCertifiedTenantAttribute,
+  getMockDeclaredTenantAttribute,
+  getMockAgreement,
+  randomArrayItem,
+} from "pagopa-interop-commons-test";
+import { genericLogger } from "pagopa-interop-commons/dist";
+import {
+  AgreementId,
+  EServiceId,
+  DescriptorId,
+  TenantId,
+  AgreementV2,
+  AgreementAddedV2,
+  toAgreementStateV2,
+  agreementState,
+  generateId,
+  AttributeId,
+  unsafeBrandId,
+  Tenant,
+  EServiceAttribute,
+  TenantAttribute,
+  Descriptor,
+  descriptorState,
+} from "pagopa-interop-models/dist";
+import { describe, expect, it } from "vitest";
+import {
+  eServiceNotFound,
+  notLatestEServiceDescriptor,
+  descriptorNotInExpectedState,
+  agreementAlreadyExists,
+  tenantNotFound,
+  missingCertifiedAttributesError,
+} from "../src/model/domain/errors";
+import { agreementCreationConflictingStates } from "../src/model/domain/validators";
+import { ApiAgreementPayload } from "../src/model/types";
+import {
+  readLastAgreementEvent,
+  addOneEService,
+  addOneTenant,
+  agreementService,
+  addOneAgreement,
+} from "./utils";
+
+/**
+ * Executes the generic agreement expectation for agreement creation process,
+ * and return the created AgreementV1 object to be used for further checks.
+ *
+ * @param agreementId - The ID of the agreement.
+ * @param expectedEserviceId - The expected e-service ID of the agreement.
+ * @param expectedDescriptorId - The expected descriptor ID of the agreement.
+ * @param expectedProducerId - The expected producer ID of the agreement.
+ * @param expectedConsumerId - The expected consumer ID of the agreement.
+ * @returns A Promise that resolves return the created AgreementV1 object.
+ */
+const expectedAgreementCreation = async (
+  agreementId: AgreementId | undefined,
+  expectedEserviceId: EServiceId,
+  expectedDescriptorId: DescriptorId,
+  expectedProducerId: TenantId,
+  expectedConsumerId: TenantId
+): Promise<AgreementV2> => {
+  expect(agreementId).toBeDefined();
+  if (!agreementId) {
+    fail("Unhandled error: returned agreementId is undefined");
+  }
+
+  const writtenEvent = await readLastAgreementEvent(agreementId);
+
+  if (!writtenEvent) {
+    fail("Creation fails: agreement not found in event-store");
+  }
+
+  expect(writtenEvent).toMatchObject({
+    type: "AgreementAdded",
+    event_version: 2,
+    version: "0",
+    stream_id: agreementId,
+  });
+
+  const actualAgreement: AgreementV2 | undefined = decodeProtobufPayload({
+    messageType: AgreementAddedV2,
+    payload: writtenEvent.data,
+  }).agreement;
+
+  if (!actualAgreement) {
+    fail("impossible to decode AgreementAddedV1 data");
+  }
+
+  expect(actualAgreement.contract).toBeUndefined();
+  expect(actualAgreement).property("createdAt").satisfy(expectPastTimestamp);
+
+  expect(actualAgreement).toMatchObject({
+    id: agreementId,
+    eserviceId: expectedEserviceId,
+    descriptorId: expectedDescriptorId,
+    producerId: expectedProducerId,
+    consumerId: expectedConsumerId,
+    state: toAgreementStateV2(agreementState.draft),
+    verifiedAttributes: [],
+    certifiedAttributes: [],
+    declaredAttributes: [],
+    consumerDocuments: [],
+    stamps: {},
+    createdAt: expect.any(BigInt),
+  });
+
+  return actualAgreement;
+};
+
+describe("create agreement", () => {
+  it("should succeed when EService Producer and Agreement Consumer are the same, even on unmet attributes", async () => {
+    const authData = getRandomAuthData();
+    const eserviceId = generateId<EServiceId>();
+    const descriptorId = generateId<DescriptorId>();
+    const attributeId = generateId<AttributeId>();
+
+    const descriptor = getMockDescriptorPublished(descriptorId, [
+      [getMockEServiceAttribute(attributeId)],
+    ]);
+    const eservice = getMockEService(eserviceId, authData.organizationId, [
+      descriptor,
+    ]);
+    const tenant = getMockTenant(authData.organizationId);
+
+    await addOneEService(eservice);
+    await addOneTenant(tenant);
+
+    const agreementData: ApiAgreementPayload = {
+      eserviceId,
+      descriptorId,
+    };
+    const createdAgreement = await agreementService.createAgreement(
+      agreementData,
+      { authData, correlationId: "", serviceName: "", logger: genericLogger }
+    );
+
+    await expectedAgreementCreation(
+      unsafeBrandId<AgreementId>(createdAgreement.id),
+      eserviceId,
+      descriptorId,
+      authData.organizationId,
+      tenant.id
+    );
+  });
+
+  it("should succeed when EService producer and Agreement consumer are different Tenants, and the consumer has all Descriptor certified Attributes not revoked", async () => {
+    const authData = getRandomAuthData();
+    const eserviceProducer: Tenant = getMockTenant();
+
+    const certifiedDescriptorAttribute1: EServiceAttribute =
+      getMockEServiceAttribute();
+    const certifiedDescriptorAttribute2: EServiceAttribute =
+      getMockEServiceAttribute();
+
+    const descriptor = getMockDescriptorPublished(generateId<DescriptorId>(), [
+      [certifiedDescriptorAttribute1],
+      [certifiedDescriptorAttribute2],
+    ]);
+
+    const certifiedTenantAttribute1: TenantAttribute = {
+      ...getMockCertifiedTenantAttribute(certifiedDescriptorAttribute1.id),
+      revocationTimestamp: undefined,
+    };
+
+    const certifiedTenantAttribute2: TenantAttribute = {
+      ...getMockCertifiedTenantAttribute(certifiedDescriptorAttribute2.id),
+      revocationTimestamp: undefined,
+    };
+
+    const consumer = getMockTenant(authData.organizationId, [
+      getMockDeclaredTenantAttribute(),
+      certifiedTenantAttribute1,
+      certifiedTenantAttribute2,
+    ]);
+
+    const eservice = getMockEService(
+      generateId<EServiceId>(),
+      eserviceProducer.id,
+      [descriptor]
+    );
+
+    await addOneTenant(eserviceProducer);
+    await addOneTenant(consumer);
+    await addOneEService(eservice);
+
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: eservice.descriptors[0].id,
+    };
+
+    const createdAgreement = await agreementService.createAgreement(
+      apiAgreementPayload,
+      { authData, correlationId: "", serviceName: "", logger: genericLogger }
+    );
+
+    await expectedAgreementCreation(
+      unsafeBrandId<AgreementId>(createdAgreement.id),
+      eservice.id,
+      descriptor.id,
+      eserviceProducer.id,
+      consumer.id
+    );
+  });
+
+  it("should succeed when EService producer and Agreement consumer are different Tenants, and the Descriptor has no certified Attributes", async () => {
+    const eserviceProducer: Tenant = getMockTenant();
+    const consumer: Tenant = getMockTenant();
+
+    // Descriptor has no certified attributes - no requirements for the consumer
+    const descriptor = getMockDescriptorPublished();
+
+    const eservice = getMockEService(
+      generateId<EServiceId>(),
+      eserviceProducer.id,
+      [descriptor]
+    );
+
+    await addOneTenant(eserviceProducer);
+    await addOneTenant(consumer);
+    await addOneEService(eservice);
+
+    const authData = getRandomAuthData(consumer.id); // different from eserviceProducer
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: eservice.descriptors[0].id,
+    };
+
+    const createdAgreement = await agreementService.createAgreement(
+      apiAgreementPayload,
+      { authData, correlationId: "", serviceName: "", logger: genericLogger }
+    );
+
+    await expectedAgreementCreation(
+      unsafeBrandId<AgreementId>(createdAgreement.id),
+      eservice.id,
+      descriptor.id,
+      eserviceProducer.id,
+      consumer.id
+    );
+  });
+
+  it("should succeed when EService's latest Descriptors are draft, and the latest non-draft Descriptor is published", async () => {
+    const tenant: Tenant = getMockTenant();
+
+    const descriptor0: Descriptor = getMockDescriptorPublished();
+    const descriptor1: Descriptor = {
+      ...getMockDescriptorPublished(),
+      version: "1",
+      state: descriptorState.draft,
+    };
+
+    const descriptor2: Descriptor = {
+      ...getMockDescriptorPublished(),
+      version: "2",
+      state: descriptorState.draft,
+    };
+
+    const eservice = getMockEService(generateId<EServiceId>(), tenant.id, [
+      descriptor0,
+      descriptor1,
+      descriptor2,
+    ]);
+
+    await addOneTenant(tenant);
+    await addOneEService(eservice);
+
+    const authData = getRandomAuthData(tenant.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: descriptor0.id,
+    };
+
+    const createdAgreement = await agreementService.createAgreement(
+      apiAgreementPayload,
+      { authData, correlationId: "", serviceName: "", logger: genericLogger }
+    );
+
+    await expectedAgreementCreation(
+      unsafeBrandId<AgreementId>(createdAgreement.id),
+      eservice.id,
+      descriptor0.id,
+      tenant.id,
+      tenant.id
+    );
+  });
+
+  it("should succeed when Agreements in non-conflicting states exist for the same EService and consumer", async () => {
+    const tenant: Tenant = getMockTenant();
+    const descriptor: Descriptor = getMockDescriptorPublished();
+
+    const eservice = getMockEService(generateId<EServiceId>(), tenant.id, [
+      descriptor,
+    ]);
+
+    const otherAgreement = getMockAgreement(
+      eservice.id,
+      tenant.id,
+      randomArrayItem(
+        Object.values(agreementState).filter(
+          (state) => !agreementCreationConflictingStates.includes(state)
+        )
+      )
+    );
+
+    await addOneTenant(tenant);
+    await addOneEService(eservice);
+    await addOneAgreement(otherAgreement);
+
+    const authData = getRandomAuthData(tenant.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: descriptor.id,
+    };
+
+    const createdAgreement = await agreementService.createAgreement(
+      apiAgreementPayload,
+      { authData, correlationId: "", serviceName: "", logger: genericLogger }
+    );
+
+    await expectedAgreementCreation(
+      unsafeBrandId<AgreementId>(createdAgreement.id),
+      eservice.id,
+      descriptor.id,
+      tenant.id,
+      tenant.id
+    );
+  });
+
+  it("should throw an eServiceNotFound error when the EService does not exist", async () => {
+    const authData = getRandomAuthData();
+    const eserviceId = generateId<EServiceId>();
+    const descriptorId = generateId<DescriptorId>();
+
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId,
+      descriptorId,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      eServiceNotFound(unsafeBrandId(apiAgreementPayload.eserviceId))
+    );
+  });
+
+  it("should throw a notLatestEServiceDescriptor error when the EService has no Descriptor", async () => {
+    const authData = getRandomAuthData();
+    const eserviceId = generateId<EServiceId>();
+
+    const eservice = getMockEService(eserviceId, authData.organizationId, []);
+
+    await addOneEService(eservice);
+
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId,
+      descriptorId: generateId<DescriptorId>(),
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      notLatestEServiceDescriptor(
+        unsafeBrandId(apiAgreementPayload.descriptorId)
+      )
+    );
+  });
+
+  it("should throw a notLatestEServiceDescriptor error when the EService Descriptor is not the latest non-draft Descriptor", async () => {
+    const authData = getRandomAuthData();
+    const eserviceId = generateId<EServiceId>();
+    const notDraftDescriptorStates = Object.values(descriptorState).filter(
+      (state) => state !== descriptorState.draft
+    );
+
+    const descriptor0: Descriptor = {
+      ...getMockDescriptorPublished(),
+      version: "0",
+      state: randomArrayItem(notDraftDescriptorStates),
+    };
+    const descriptor1: Descriptor = {
+      ...getMockDescriptorPublished(),
+      version: "1",
+      state: randomArrayItem(notDraftDescriptorStates),
+    };
+
+    const eservice = getMockEService(eserviceId, authData.organizationId, [
+      descriptor0,
+      descriptor1,
+    ]);
+
+    await addOneEService(eservice);
+    await addOneTenant(getMockTenant(authData.organizationId));
+
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId,
+      descriptorId: descriptor0.id,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      notLatestEServiceDescriptor(
+        unsafeBrandId(apiAgreementPayload.descriptorId)
+      )
+    );
+  });
+
+  it("should throw a descriptorNotInExpectedState error when the EService's latest non-draft Descriptor is not published", async () => {
+    const authData = getRandomAuthData();
+    const eserviceId = generateId<EServiceId>();
+
+    const descriptor: Descriptor = {
+      ...getMockDescriptorPublished(),
+      version: "0",
+      state: randomArrayItem(
+        Object.values(descriptorState).filter(
+          (state) =>
+            state !== descriptorState.published &&
+            state !== descriptorState.draft
+        )
+      ),
+    };
+
+    const eservice = getMockEService(eserviceId, authData.organizationId, [
+      descriptor,
+    ]);
+
+    await addOneEService(eservice);
+    await addOneTenant(getMockTenant(authData.organizationId));
+
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId,
+      descriptorId: descriptor.id,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      descriptorNotInExpectedState(eservice.id, descriptor.id, [
+        descriptorState.published,
+      ])
+    );
+  });
+
+  it("should throw an agreementAlreadyExists error when an Agreement in a conflicting state already exists for the same EService and consumer", async () => {
+    const consumer: Tenant = getMockTenant();
+    const descriptor: Descriptor = getMockDescriptorPublished();
+
+    const eservice = getMockEService(generateId<EServiceId>(), consumer.id, [
+      descriptor,
+    ]);
+
+    const conflictingAgreement = getMockAgreement(
+      eservice.id,
+      consumer.id,
+      randomArrayItem(agreementCreationConflictingStates)
+    );
+
+    await addOneTenant(consumer);
+    await addOneEService(eservice);
+    await addOneAgreement(conflictingAgreement);
+
+    const authData = getRandomAuthData(consumer.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: descriptor.id,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(agreementAlreadyExists(consumer.id, eservice.id));
+  });
+
+  it("should throw a tenantNotFound error when the consumer Tenant does not exist", async () => {
+    const consumer: Tenant = getMockTenant();
+    const descriptor: Descriptor = getMockDescriptorPublished();
+
+    const eservice = getMockEService(
+      generateId<EServiceId>(),
+      generateId<TenantId>(),
+      [descriptor]
+    );
+
+    await addOneEService(eservice);
+
+    const authData = getRandomAuthData(consumer.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: descriptor.id,
+    };
+
+    await expect(() =>
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(tenantNotFound(consumer.id));
+  });
+
+  it("should throw a missingCertifiedAttributesError error when the EService producer and Agreement consumer are different Tenants, and the consumer is missing a Descriptor certified Attribute", async () => {
+    const eserviceProducer: Tenant = getMockTenant();
+
+    // Descriptor has two certified attributes
+    const certifiedDescriptorAttribute1: EServiceAttribute =
+      getMockEServiceAttribute();
+    const certifiedDescriptorAttribute2: EServiceAttribute =
+      getMockEServiceAttribute();
+
+    const descriptor = getMockDescriptorPublished(generateId<DescriptorId>(), [
+      [certifiedDescriptorAttribute1],
+      [certifiedDescriptorAttribute2],
+    ]);
+
+    // In this case, the consumer is missing one of the two certified attributes
+    const certifiedTenantAttribute1: TenantAttribute =
+      getMockCertifiedTenantAttribute(certifiedDescriptorAttribute1.id);
+
+    const consumer = {
+      ...getMockTenant(),
+      attributes: [certifiedTenantAttribute1],
+    };
+
+    const eservice = getMockEService(
+      generateId<EServiceId>(),
+      eserviceProducer.id,
+      [descriptor]
+    );
+
+    await addOneTenant(eserviceProducer);
+    await addOneTenant(consumer);
+    await addOneEService(eservice);
+
+    const authData = getRandomAuthData(consumer.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: eservice.descriptors[0].id,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      missingCertifiedAttributesError(descriptor.id, consumer.id)
+    );
+  });
+
+  it("should throw a missingCertifiedAttributesError error when the EService producer and Agreement consumer are different Tenants, and the consumer has a Descriptor certified Attribute revoked", async () => {
+    const eserviceProducer: Tenant = getMockTenant();
+
+    // Descriptor has two certified attributes
+    const certifiedDescriptorAttribute1: EServiceAttribute =
+      getMockEServiceAttribute();
+    const certifiedDescriptorAttribute2: EServiceAttribute =
+      getMockEServiceAttribute();
+
+    const descriptor: Descriptor = getMockDescriptorPublished(
+      generateId<DescriptorId>(),
+      [[certifiedDescriptorAttribute1], [certifiedDescriptorAttribute2]]
+    );
+
+    const eservice = getMockEService(
+      generateId<EServiceId>(),
+      eserviceProducer.id,
+      [descriptor]
+    );
+
+    // In this case, the consumer has one of the two certified attributes revoked
+    const certifiedTenantAttribute1: TenantAttribute = {
+      ...getMockCertifiedTenantAttribute(certifiedDescriptorAttribute1.id),
+      revocationTimestamp: new Date(),
+      assignmentTimestamp: new Date(),
+    };
+    const certifiedTenantAttribute2: TenantAttribute = {
+      ...getMockCertifiedTenantAttribute(certifiedDescriptorAttribute2.id),
+      revocationTimestamp: undefined,
+      assignmentTimestamp: new Date(),
+    };
+
+    const consumer = {
+      ...getMockTenant(),
+      attributes: [certifiedTenantAttribute1, certifiedTenantAttribute2],
+    };
+
+    await addOneTenant(eserviceProducer);
+    await addOneTenant(consumer);
+    await addOneEService(eservice);
+
+    const authData = getRandomAuthData(consumer.id);
+    const apiAgreementPayload: ApiAgreementPayload = {
+      eserviceId: eservice.id,
+      descriptorId: eservice.descriptors[0].id,
+    };
+
+    await expect(
+      agreementService.createAgreement(apiAgreementPayload, {
+        authData,
+        correlationId: "",
+        serviceName: "",
+        logger: genericLogger,
+      })
+    ).rejects.toThrowError(
+      missingCertifiedAttributesError(descriptor.id, consumer.id)
+    );
+  });
+});
